// src/views/products/types/ProductUpdate.ts
export interface ProductUpdateVariables {
  id: string;
  input: ProductInput;
  firstValues?: number | null;
  afterValues?: string | null;
  lastValues?: number | null;
  beforeValues?: string | null;
}