// src/views/products/types/SimpleProductUpdate.ts
export interface SimpleProductUpdate_productUpdate_errors {
  __typename: "ProductError";
  code: ProductErrorCode;
  field: string | null;
  attributes: string[] | null;
}
// ...
export interface SimpleProductUpdate_productUpdate_product {
  __typename: "Product";
  id: string;
  attributes: SimpleProductUpdate_productUpdate_product_attributes[];
  productType: SimpleProductUpdate_productUpdate_product_productType;
  channelListings: SimpleProductUpdate_productUpdate_product_channelListings[] | null;
  metadata: (SimpleProductUpdate_productUpdate_product_metadata | null)[];
  privateMetadata: (SimpleProductUpdate_productUpdate_product_privateMetadata | null)[];
  name: string;
  slug: string;
  description: any | null;
  seoTitle: string | null;
  seoDescription: string | null;
  rating: number | null;
  defaultVariant: SimpleProductUpdate_productUpdate_product_defaultVariant | null;
  category: SimpleProductUpdate_productUpdate_product_category | null;
  collections: (SimpleProductUpdate_productUpdate_product_collections | null)[] | null;
  chargeTaxes: boolean;
  media: SimpleProductUpdate_productUpdate_product_media[] | null;
  isAvailable: boolean | null;
  variants: (SimpleProductUpdate_productUpdate_product_variants | null)[] | null;
  weight: SimpleProductUpdate_productUpdate_product_weight | null;
  taxType: SimpleProductUpdate_productUpdate_product_taxType | null;
}

export interface SimpleProductUpdate_productUpdate {
  __typename: "ProductUpdate";
  errors: SimpleProductUpdate_productUpdate_errors[];
  product: SimpleProductUpdate_productUpdate_product | null;
}

export interface SimpleProductUpdate_productVariantUpdate_errors {
  __typename: "ProductError";
  code: ProductErrorCode;
  field: string | null;
  attributes: string[] | null;
}
// ...
export interface SimpleProductUpdate_productVariantUpdate_productVariant {
  __typename: "ProductVariant";
  id: string;
  metadata: (SimpleProductUpdate_productVariantUpdate_productVariant_metadata | null)[];
  privateMetadata: (SimpleProductUpdate_productVariantUpdate_productVariant_privateMetadata | null)[];
  selectionAttributes: SimpleProductUpdate_productVariantUpdate_productVariant_selectionAttributes[];
  nonSelectionAttributes: SimpleProductUpdate_productVariantUpdate_productVariant_nonSelectionAttributes[];
  media: SimpleProductUpdate_productVariantUpdate_productVariant_media[] | null;
  name: string;
  product: SimpleProductUpdate_productVariantUpdate_productVariant_product;
  channelListings: SimpleProductUpdate_productVariantUpdate_productVariant_channelListings[] | null;
  sku: string;
  stocks: (SimpleProductUpdate_productVariantUpdate_productVariant_stocks | null)[] | null;
  trackInventory: boolean;
  weight: SimpleProductUpdate_productVariantUpdate_productVariant_weight | null;
}

export interface SimpleProductUpdate_productVariantUpdate {
  __typename: "ProductVariantUpdate";
  errors: SimpleProductUpdate_productVariantUpdate_errors[];
  productVariant: SimpleProductUpdate_productVariantUpdate_productVariant | null;
}

export interface SimpleProductUpdate_productVariantStocksCreate_errors {
  __typename: "BulkStockError";
  code: ProductErrorCode;
  field: string | null;
  index: number | null;
}
// ...
export interface SimpleProductUpdate_productVariantStocksCreate_productVariant {
  __typename: "ProductVariant";
  id: string;
  metadata: (SimpleProductUpdate_productVariantStocksCreate_productVariant_metadata | null)[];
  privateMetadata: (SimpleProductUpdate_productVariantStocksCreate_productVariant_privateMetadata | null)[];
  selectionAttributes: SimpleProductUpdate_productVariantStocksCreate_productVariant_selectionAttributes[];
  nonSelectionAttributes: SimpleProductUpdate_productVariantStocksCreate_productVariant_nonSelectionAttributes[];
  media: SimpleProductUpdate_productVariantStocksCreate_productVariant_media[] | null;
  name: string;
  product: SimpleProductUpdate_productVariantStocksCreate_productVariant_product;
  channelListings: SimpleProductUpdate_productVariantStocksCreate_productVariant_channelListings[] | null;
  sku: string;
  stocks: (SimpleProductUpdate_productVariantStocksCreate_productVariant_stocks | null)[] | null;
  trackInventory: boolean;
  weight: SimpleProductUpdate_productVariantStocksCreate_productVariant_weight | null;
}

export interface SimpleProductUpdate_productVariantStocksCreate {
  __typename: "ProductVariantStocksCreate";
  errors: SimpleProductUpdate_productVariantStocksCreate_errors[];
  productVariant: SimpleProductUpdate_productVariantStocksCreate_productVariant | null;
}

export interface SimpleProductUpdate_productVariantStocksDelete_errors {
  __typename: "StockError";
  code: StockErrorCode;
  field: string | null;
}
// ...
export interface SimpleProductUpdate_productVariantStocksDelete_productVariant {
  __typename: "ProductVariant";
  id: string;
  metadata: (SimpleProductUpdate_productVariantStocksDelete_productVariant_metadata | null)[];
  privateMetadata: (SimpleProductUpdate_productVariantStocksDelete_productVariant_privateMetadata | null)[];
  selectionAttributes: SimpleProductUpdate_productVariantStocksDelete_productVariant_selectionAttributes[];
  nonSelectionAttributes: SimpleProductUpdate_productVariantStocksDelete_productVariant_nonSelectionAttributes[];
  media: SimpleProductUpdate_productVariantStocksDelete_productVariant_media[] | null;
  name: string;
  product: SimpleProductUpdate_productVariantStocksDelete_productVariant_product;
  channelListings: SimpleProductUpdate_productVariantStocksDelete_productVariant_channelListings[] | null;
  sku: string;
  stocks: (SimpleProductUpdate_productVariantStocksDelete_productVariant_stocks | null)[] | null;
  trackInventory: boolean;
  weight: SimpleProductUpdate_productVariantStocksDelete_productVariant_weight | null;
}

export interface SimpleProductUpdate_productVariantStocksDelete {
  __typename: "ProductVariantStocksDelete";
  errors: SimpleProductUpdate_productVariantStocksDelete_errors[];
  productVariant: SimpleProductUpdate_productVariantStocksDelete_productVariant | null;
}

export interface SimpleProductUpdate_productVariantStocksUpdate_errors {
  __typename: "BulkStockError";
  code: ProductErrorCode;
  field: string | null;
  index: number | null;
}
// ...
export interface SimpleProductUpdate_productVariantStocksUpdate_productVariant {
  __typename: "ProductVariant";
  id: string;
  metadata: (SimpleProductUpdate_productVariantStocksUpdate_productVariant_metadata | null)[];
  privateMetadata: (SimpleProductUpdate_productVariantStocksUpdate_productVariant_privateMetadata | null)[];
  selectionAttributes: SimpleProductUpdate_productVariantStocksUpdate_productVariant_selectionAttributes[];
  nonSelectionAttributes: SimpleProductUpdate_productVariantStocksUpdate_productVariant_nonSelectionAttributes[];
  media: SimpleProductUpdate_productVariantStocksUpdate_productVariant_media[] | null;
  name: string;
  product: SimpleProductUpdate_productVariantStocksUpdate_productVariant_product;
  channelListings: SimpleProductUpdate_productVariantStocksUpdate_productVariant_channelListings[] | null;
  sku: string;
  stocks: (SimpleProductUpdate_productVariantStocksUpdate_productVariant_stocks | null)[] | null;
  trackInventory: boolean;
  weight: SimpleProductUpdate_productVariantStocksUpdate_productVariant_weight | null;
}

export interface SimpleProductUpdate_productVariantStocksUpdate {
  __typename: "ProductVariantStocksUpdate";
  errors: SimpleProductUpdate_productVariantStocksUpdate_errors[];
  productVariant: SimpleProductUpdate_productVariantStocksUpdate_productVariant | null;
}

export interface SimpleProductUpdate {
  productUpdate: SimpleProductUpdate_productUpdate | null;
  productVariantUpdate: SimpleProductUpdate_productVariantUpdate | null;
  productVariantStocksCreate: SimpleProductUpdate_productVariantStocksCreate | null;
  productVariantStocksDelete: SimpleProductUpdate_productVariantStocksDelete | null;
  productVariantStocksUpdate: SimpleProductUpdate_productVariantStocksUpdate | null;
}